implement draw_sprite instruction
- converted internal Display.pixels representation to 2D Vec
  of booleans for easier drawing
- public APIs to get display pixels is Vec<Vec<u8>> (row, byte col)
- sprites are represented as Vec<u8> where 0 < len < 16

+++ src/chip8/registers.rs
@@ -1,6 +1,6 @@
 pub struct Registers {
     pub v: Vec<u8>,
-    pub i: u16,
+    pub i: usize,
     pub pc: u16,
     pub sp: usize,
     pub stack: Vec<u16>,
